another type thing

+++ Drivers/SSD168x/Source/Ssd168xDisplay.cpp
@@ -223,7 +223,7 @@ void Ssd168xDisplay::lvglFlushCallback(lv_display_t* disp, const lv_area_t* area
         }
     }
 
-    const int clamped_x = std::max(area->x1, 0);
+    const int clamped_x = std::max(area->x1, static_cast<int32_t>(0));
     const int clamped_y = std::max(area->y1, static_cast<int32_t>(0));
     const int max_w = std::max(0, panel_width - clamped_x);
     const int max_h = std::max(0, panel_height - clamped_y);
